Fixed crash bug where VkDebugUtilsMessengerEXT try to get its destroy procAddr during desturction with a nullptr instance

+++ include/VulkanObjectWrapper.h
@@ -141,9 +141,11 @@ public:
 
 	~InnerHelper()
 	{
-	    auto destroyFunc = reinterpret_cast<DestroyFuncType>(
-		vkGetInstanceProcAddr(this->destroyDependency, DestroyFuncName));
-	    destroyFunc(this->destroyDependency, this->handle, nullptr);
+	    if (this->destroyDependency != nullptr) {
+		auto destroyFunc = reinterpret_cast<DestroyFuncType>(
+		    vkGetInstanceProcAddr(this->destroyDependency, DestroyFuncName));
+		destroyFunc(this->destroyDependency, this->handle, nullptr);
+	    }
 	}
 
 	InnerHelper(InnerHelper&&) = default;
